Basic Webpage complete
All of the requirements were met for assignement 1 and are incorporated into the html.index file that is being updated now. There is one bug that I was able to find but don't have time to fix yet; when the user clicks no "Previous Courses" or "Experience", the popup circle containing the information is not centered in the screen. Additionally for some reason, the formatting of the "Previous Courses" inside the list view is not formatting as it would be expected. This may be fixed after the submission, time permitting.